pyproject.toml
[project]
name = "jupiter-iaa-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb==1.2.1",
    "fastmcp>=2.0",
    "gradio>=5.44.1",
    "gtts>=2.5.4",
    "hf-xet>=1.1.9",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=3.0.0",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "langchain-qdrant>=1.1.0",
    "langgraph>=0.6.6",
    "langgraph-api>=0.4.7",
    "langgraph-cli>=0.4.0",
    "notebook>=7.4.5",
    "openai>=1.102.0",
    "openai-agents>=0.2.10",
    "plotly>=6.3.0",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
]
